backend/requirements.txt

Django==5.0.7 
djangorestframework==3.16.1
django-cors-headers==4.4.0
daphne==4.1.2
psycopg==3.2.10
psycopg-binary==3.2.10
redis==5.0.7
scikit-learn==1.5.1
python-dotenv==1.0.1
pdfminer.six==20231228
markdown-it-py==3.0.0
python-jose[cryptography]==3.3.0
requests==2.32.3
cachetools==5.3.3
tiktoken>=0.7.0
google-generativeai>=0.8.0
django-tenants==3.6.1
channels==4.1.0
channels-redis==4.2.0
pgvector==0.2.5
PyJWT==2.9.0
python-keycloak==4.1.0
psycopg[binary]==3.2.10
rank-bm25==0.2.2
prometheus-client==0.21.0
django-redis==5.4.0
ujson==5.10.0
structlog==24.1.0